add task 307 and fix task 3

+++ L303/script.js
@@ -1,6 +1,6 @@
 var input = prompt('შეიტანეთ თანხა', '$10');
 var isDollar = input.indexOf('$') > -1;
-var amount = parseInt(input.substring(1, input.length));
+var amount = parseFloat(input.substring(1, input.length));
 if (isDollar) {
     alert(input + '=L' + amount * 2.6);
 }else {
